IP 1.9.3: Added Materials elements, Did registry form, did NavBar, stylized.
