+++ gym.frontend/client-app/src/app/app.component.html
@@ -1,3 +1,5 @@
+<app-navbar></app-navbar>
+
 <div>
 <p>Hello, this server works!</p>
 <button mat-fab color="primary" aria-label="Example icon button with a delete icon">
@@ -5,3 +7,5 @@
   </button>
 <div>{{loggedUser?.id}} {{loggedUser?.firstName}} {{loggedUser?.lastName}} {{loggedUser?.username}}</div>
 </div>
+
+<app-register-form></app-register-form>
