+++ gym.frontend/client-app/src/app/app.component.ts
@@ -7,13 +7,16 @@ import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
+import { NavbarComponent } from './home/navbar/navbar/navbar.component';
+import { MaterialModule } from './material/material.module';
+import { RegisterFormComponent } from './register.form/register/register.form.component';
 
 @Component({
   selector: 'app-root',
   standalone: true,
   templateUrl: './app.component.html',
   styleUrl: './app.component.css',
-  imports: [HttpClientModule, MatButtonModule, MatIconModule],
+  imports: [HttpClientModule, MaterialModule, NavbarComponent, RegisterFormComponent],
   providers: [UserService]
 })
 export class AppComponent implements OnInit{
