+++ gym.frontend/client-app/src/app/home/navbar/navbar/navbar.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { MaterialModule } from '../../../material/material.module';
+
+@Component({
+  selector: 'app-navbar',
+  standalone: true,
+  imports: [MaterialModule],
+  templateUrl: './navbar.component.html',
+  styleUrl: './navbar.component.css'
+})
+export class NavbarComponent {
+
+}
