+++ gym.frontend/client-app/src/app/material/material.module.ts
@@ -1,9 +1,17 @@
 import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import {MatInputModule} from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
 
 
 const MaterialComponents = [
-  MatButtonModule
+  MatButtonModule,
+  MatIconModule,
+  MatMenuModule,
+  MatInputModule,
+  MatFormFieldModule
 ];
 
 @NgModule({
