--- gym.frontend/client-app/src/app/register.form/register.form.component.html
@@ -1,1 +1,0 @@
-<p>register.form works!</p>
