+++ gym.frontend/client-app/src/app/register.form/register/register.form.component.css
@@ -1,0 +1,9 @@
+.example-form {
+    min-width: 150px;
+    max-width: 500px;
+    width: 100%;
+  }
+  
+  .example-full-width {
+    width: 100%;
+  }
