+++ gym.frontend/client-app/src/app/register.form/register/register.form.component.ts
@@ -1,0 +1,28 @@
+import { Component } from '@angular/core';
+import { FormControl, FormGroupDirective, FormsModule, NgForm, ReactiveFormsModule, Validators } from '@angular/forms';
+import { MaterialModule } from '../../material/material.module';
+import { ErrorStateMatcher } from '@angular/material/core';
+
+@Component({
+  selector: 'app-register-form',
+  standalone: true,
+  imports: [MaterialModule, FormsModule, ReactiveFormsModule],
+  templateUrl: './register.form.component.html',
+  styleUrl: './register.form.component.css'
+})
+export class RegisterFormComponent {
+  emailFormControl = new FormControl('', [Validators.required, Validators.email]);
+
+  matcher = new MyErrorStateMatcher();
+}
+
+export class MyErrorStateMatcher implements ErrorStateMatcher {
+  isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+    const isSubmitted = form && form.submitted;
+    return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
+  }
+
+
+}
+
+
